fixed warnings

+++ ios/SmileAndFrowns/SNFAddSmileOrFrown.m
@@ -15,7 +15,7 @@
 	_smileCountPicker = [[SNFValuePicker alloc] init];
 	NSMutableArray *vals = [[NSMutableArray alloc] init];
 	for(NSInteger i=1; i<101; i++){
-		[vals addObject:[NSString stringWithFormat:@"%lu", i]];
+		[vals addObject:[NSString stringWithFormat:@"%lu", (long)i]];
 	}
 	_smileCountPicker.values = vals;
 	_smileCountPicker.delegate = self;
@@ -225,7 +225,7 @@
 - (IBAction)onChangeAmount:(UIButton *)sender{
 	[self.view addSubview:_smileCountPicker.view];
 	[_smileCountPicker.view  matchFrameSizeOfView:self.view];
-	_smileCountPicker.selectedValue = [NSString stringWithFormat:@"%lu", (NSInteger)floor(self.amountStepper.value)];
+	_smileCountPicker.selectedValue = [NSString stringWithFormat:@"%lu", (long)floor(self.amountStepper.value)];
 }
 
 - (void)valuePicker:(SNFValuePicker *)valuePicker changedValue:(NSString *)value{
